requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.1
colorama==0.4.4
cryptography==36.0.0
dnspython==1.16.0
ecdsa==0.17.0
fastapi==0.68.1
Flask==2.0.1
Flask-PyMongo==2.3.0
h11==0.12.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pymongo==3.12.0
python-jose==3.3.0
rsa==4.8
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
Werkzeug==2.0.1
